refactor: extract scenario exporting functionality to scenario/exporter_csv.py
- Moved all CSV-export helper methods and file IO operations out of world.py into a new dedicated module assume/scenario/exporter_csv.py
- Replaced the inline exporting logic in World.export with a delegation to assume.scenario.exporter_csv.export_to_folder
- Cleaned up unused imports (yaml, timedelta) in world.py
